MAINT: move modules to package hierarchy
Modules pars and config
* Update the imports of these modules in other files
* Actually delete the import of config, which mostly alters the path
