+++ src/flexfrac1d/pars.py
